+++ Source/EventFlow/ValueObjects/SingleValueObject.cs
@@ -25,12 +25,10 @@ using System.Collections.Generic;
 
 namespace EventFlow.ValueObjects
 {
-    public abstract class SingleValueObject<T> : ValueObject, IComparable
+    public abstract class SingleValueObject<T> : ValueObject, IComparable, ISingleValueObject
         where T : IComparable, IComparable<T>
     {
         public T Value { get; private set; }
-
-        protected SingleValueObject() { }
 
         protected SingleValueObject(T value)
         {
@@ -66,6 +64,11 @@ namespace EventFlow.ValueObjects
             return ReferenceEquals(Value, null)
                 ? string.Empty
                 : Value.ToString();
+        }
+
+        public object GetValue()
+        {
+            return Value;
         }
     }
 }